soc/amd/stoneyridge: Use uint8_t as type for SPD address
SPD address is currenty int. It should be uint8_t.

BUG=b:62200225

Reviewed-on: https://review.coreboot.org/22300
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/soc/amd/common/dimmSpd.h
@@ -20,7 +20,7 @@ AGESA_STATUS
 AmdMemoryReadSPD(IN UINT32 Func, IN UINTN Data,
 				IN OUT AGESA_READ_SPD_PARAMS *SpdData);
 
-int sb_readSpd(int spdAddress, char *buf, size_t len);
+int sb_readSpd(uint8_t spdAddress, char *buf, size_t len);
 int smbus_readSpd(int spdAddress, char *buf, size_t len);
 
 #endif
